refactor: Move QuadIndexPattern and RingAllocator to Core
