+++ src/CutTheRopeDX.Desktop/Desktop/QuadBatch.cs
@@ -1,5 +1,7 @@
 using System;
 
+using CutTheRopeDX.Framework.Platform;
+
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
